requirements.txt
-e git+https://github.com/labscript-suite/blacs@511975b6a1f2b280f7df6b1021a9ec616aee57e4#egg=blacs
contourpy==1.3.1
cycler==0.12.1
desktop-app==0.4.1
distlib==0.3.9
fonttools==4.57.0
h5py==3.13.0
importlib_metadata==8.6.1
kiwisolver==1.4.8
-e git+https://github.com/labscript-suite/labscript@bfc65febc417c0e6ae9c7ad591f9a4d52fcebe14#egg=labscript
labscript-c-extensions==1.2.0
-e git+https://github.com/labscript-suite/labscript-devices@5e71c14482a22e49ca0f1838d183652278cfc6e1#egg=labscript_devices
-e git+https://github.com/labscript-suite/labscript-utils@8c11c95a6db1616be9feb87dc6321a0c2a808ccb#egg=labscript_utils
libusb1==3.3.1
-e git+https://github.com/labscript-suite/lyse@2ff743f4b764e7b074dd24aa8bf49b8da9f4b30e#egg=lyse
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.51
PyDAQmx==1.4.7
pynivision==2015.0.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
PyVISA==1.15.0
pyzmq==26.4.0
qtutils==4.0.0
-e git+https://github.com/labscript-suite/runmanager@c66c8df8fc67912a0612deef49953d5b874150e4#egg=runmanager
-e git+https://github.com/labscript-suite/runviewer@1f9fc063975a5e3e9935d04c78dcbb31992fa05a#egg=runviewer
scipy==1.15.2
setuptools==78.1.0
setuptools-scm==8.2.0
six==1.17.0
spinapi @ file:///home/localadmin_artemis/Documents/labscript-install/labscript-install/spinapi-python
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
zprocess==2.26.1
